schooloverviewpage to es6
